# Conflicts:
#	src/yeet/include/parse_args.hpp

+++ src/common/wflign/src/rkmh.hpp
@@ -40,13 +40,13 @@ static char rev_arr[26] = {
         85, 86, 87, 88, 89, 90
 };
 
-typedef uint64_t hash_t;
+typedef uint32_t hash_t;
 
 std::vector<hash_t> hash_sequence(const char* seq,
                                   const uint64_t& len,
                                   const uint64_t& k,
                                   const uint64_t& sketch_size);
 
-double compare(const std::vector<hash_t>& alpha, const std::vector<hash_t>& beta, const uint64_t& k);
+float compare(const std::vector<hash_t>& alpha, const std::vector<hash_t>& beta, const uint64_t& k);
 
 }
